+++ doc/changelog.rst
@@ -9,6 +9,11 @@ Unrealeased
   When releasing next version, the "Unreleased" header will be replaced
   with appropriate version header and this help text will be removed.
 
+Front
+~~~~~
+
+- Fixed manufacturer modified filter to consider only visible shop products
+
 Classic Gray Theme
 ~~~~~~~~~~~~~~~~~~